[refactor] 패키지 구조 수정
- 레이어드 구조

- controller, service 패키지 추가 및 분리
