+++ easytask/src/main/java/com/easytask/easytask/src/task/service/TaskService.java
@@ -1,4 +1,4 @@
-package com.easytask.easytask.src.task;
+package com.easytask.easytask.src.task.service;
 
 import com.easytask.easytask.common.exception.BaseException;
 import com.easytask.easytask.common.scheduler.MatchingRequest;
