+++ easytask/src/main/java/com/easytask/easytask/src/user/controller/UserController.java
@@ -1,8 +1,9 @@
-package com.easytask.easytask.src.user;
+package com.easytask.easytask.src.user.controller;
 
 
 import com.easytask.easytask.common.exception.BaseException;
 import com.easytask.easytask.common.response.BaseResponse;
+import com.easytask.easytask.src.user.service.UserService;
 import com.easytask.easytask.src.user.dto.request.UserLoginDto;
 import com.easytask.easytask.src.user.dto.request.UserRequestDto;
 import com.easytask.easytask.src.user.dto.request.AbilitySettingRequestDto;
