+++ easytask/src/test/java/com/easytask/easytask/src/review/ReviewRepositoryTest.java
@@ -4,8 +4,10 @@ import com.easytask.easytask.src.review.repository.ReviewRepository;
 import com.easytask.easytask.src.task.entity.RelatedAbility;
 import com.easytask.easytask.src.task.entity.Task;
 import com.easytask.easytask.src.task.entity.TaskUserMapping;
-import com.easytask.easytask.src.user.UserController;
-import com.easytask.easytask.src.user.UserService;
+import com.easytask.easytask.src.task.repository.TaskRepository;
+import com.easytask.easytask.src.user.controller.UserController;
+import com.easytask.easytask.src.user.repository.UserRepository;
+import com.easytask.easytask.src.user.service.UserService;
 import com.easytask.easytask.src.user.entity.User;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
