+++ easytask/src/test/java/com/easytask/easytask/src/task/TaskRepositoryTest.java
@@ -3,7 +3,8 @@ package com.easytask.easytask.src.task;
 import com.easytask.easytask.src.task.entity.RelatedAbility;
 import com.easytask.easytask.src.task.entity.Task;
 import com.easytask.easytask.src.task.entity.TaskUserMapping;
-import com.easytask.easytask.src.user.UserService;
+import com.easytask.easytask.src.task.repository.TaskRepository;
+import com.easytask.easytask.src.user.service.UserService;
 import com.easytask.easytask.src.user.entity.User;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
