+++ easytask/src/test/java/com/easytask/easytask/src/user/UserRepositoryTest.java
@@ -1,12 +1,14 @@
 package com.easytask.easytask.src.user;
 
+import com.easytask.easytask.src.user.controller.UserController;
 import com.easytask.easytask.src.user.entity.PossibleTask;
 import com.easytask.easytask.src.user.entity.User;
+import com.easytask.easytask.src.user.repository.UserRepository;
+import com.easytask.easytask.src.user.service.UserService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.Commit;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
